refactor: Move Task to wvlet.uni.control; drop progress API; simplify
Per review:
- Move from new wvlet.uni.concurrent (uni-core) to wvlet.uni.control
  (uni), alongside Retry / CircuitBreaker / RateLimiter / Guard /
  Resource — the existing home for lifecycle / reliability primitives.
- Consolidate Task.scala + TaskImpl.scala into a single file, matching
  Retry.scala / CircuitBreaker.scala convention.
- Drop the progress API entirely. Bodies publish progress / partial
  results / anything else through any existing Rx primitive (RxVar,
  RxQueue, RxDeferred) — the same "side-effects managed outside the
  interface" rule that already justified no `[A]` result type.
  Removes Task.Progress (type Any), progressRef, progressVar,
  progressStream, reportProgress.
- Drop Task.State.Cancelling — its own documentation acknowledged the
  transition was unreliable as an observable state. isCancelled covers
  "cancel requested"; the terminal state is the source of truth.
- Memoise awaitRx (was rebuilding the Rx chain per call).
- Inline `start` into taskCompat.run (was an unnecessary indirection).

All tests pass: JVM 9/9, JS 8/8, Native 9/9.
