+++ uni/.jvm/src/main/scala/wvlet/uni/control/taskCompat.scala
@@ -11,32 +11,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.uni.concurrent
+package wvlet.uni.control
 
 import java.util.concurrent.CountDownLatch
 import wvlet.uni.util.ThreadUtil
 
-/**
-  * Scala Native uses the same JDK threading primitives as JVM, so this mirrors the JVM impl. The
-  * only intentional difference is `Thread.interrupt()` — Scala Native supports it, but its
-  * semantics against libc blocking syscalls are weaker than HotSpot's. Bodies that need unblocking
-  * on cancel should still observe `isCancelled` at loop checkpoints.
-  */
-private[concurrent] object taskCompat:
+private[control] object taskCompat:
 
   private lazy val threadFactory = ThreadUtil.newDaemonThreadFactory("uni-task")
 
   def run(body: TaskContext => Unit): Task =
-    val task = new NativeTaskImpl()
-    task.start(body)
+    val task = new JvmTaskImpl()
+    task.scheduleBody(body)
     task
 
-  private class NativeTaskImpl extends TaskImpl:
+  private class JvmTaskImpl extends TaskImpl:
     private val latch = new CountDownLatch(1)
+    // Captured for onCancelRequested to call `interrupt()`. Set before `t.start()` runs.
     @volatile
     private var worker: Thread = null
 
-    override protected def scheduleBody(body: TaskContext => Unit): Unit =
+    override def scheduleBody(body: TaskContext => Unit): Unit =
       val t = threadFactory.newThread { () =>
         try runBody(body)
         finally latch.countDown()
@@ -44,17 +39,19 @@ private[concurrent] object taskCompat:
       worker = t
       t.start()
 
-    override protected def awaitTerminal(): Unit =
+    override def awaitTerminal(): Unit =
       latch.await()
       val cause = terminalFailure
       if cause != null then
         throw cause
 
-    override protected def onCancelRequested(): Unit =
+    override def onCancelRequested(): Unit =
       val t = worker
       if t != null then
+        // Interrupt unwedges blocking JDK calls (Thread.sleep, blocking IO, …) so the body's
+        // next checkCancelled — or the JDK call itself — exits promptly.
         t.interrupt()
 
-  end NativeTaskImpl
+  end JvmTaskImpl
 
 end taskCompat
